backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.50.0
anthropic>=0.40.0
google-genai>=1.0.0
pydantic==2.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
stripe==9.12.0
supabase>=2.10.0
